refactor(abstract-utxo): move custom change wallet tests to separate file
Move tests related to custom change wallets to their own test file for better
organization. These tests were previously part of the general abstract UTXO
coin test file.

Issue: BTC-2652
